% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{bias_corrected_est}
\alias{bias_corrected_est}
\title{bias_corrected_est}
\usage{
bias_corrected_est(est, replicates, na.rm = T)
}
\arguments{
\item{na.rm}{}
}
\value{
numeric bias corrected estimate.
}
\description{
bias_corrected_est
}
