% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{boot_ab}
\alias{boot_ab}
\title{boot_ab}
\usage{
boot_ab(x, nsim, ncpus, re.form = NA, type = c("parametric",
  "semiparametric", "cases"), how = c("groups", "within", "both"),
  seed = NULL)
}
\arguments{
\item{x}{a permediatrMod object.}

\item{re.form}{formula, NA (equivalent to \code{\link[lme4]{bootMer}}'s
\code{use.u=FALSE}; new normal deviates are drawn), or NULL (equivalent to
\code{\link[lme4]{bootMer}}'s \code{use.u=TRUE}; all inference is
conditional on these values) or formula specifying which random effects to
incorporate. See \code{\link[lme4]{simulate.merMod}} for further details.
Not used if type == 'cases'.}

\item{type}{character string specifying the type of bootstrap,
\code{"parametric"} or \code{"semiparametric"}; partial matching is
allowed.}

\item{how}{character specifying the type of resampling if type == 'cases'.}

\item{seed}{}
}
\description{
See \code{\link[lme4]{bootMer}} for more information on arguments for
\code{re.form} and \code{type}.
}
