% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{bootMerCases}
\alias{bootMerCases}
\title{bootMerCases}
\usage{
bootMerCases(x, FUN, nsim = 1, seed = NULL, how = c("groups",
  "within", "both"), parallel = c("no", "multicore"),
  ncpus = getOption("boot.ncpus", 1L))
}
\arguments{
\item{x}{a fitted \code{merMod} object: see \code{\link[lme4]{lmer}},
\code{\link[lme4]{glmer}}, etc.}

\item{FUN}{a function taking a fitted \code{merMod} object as input and
returning the statistic of interest, which must be a (possibly named)
numeric vector.}

\item{nsim}{number of resampled data sets, positive integer; the bootstrap B
(or R).}

\item{seed}{optional argument to \code{\link[base]{set.seed}}.}

\item{how}{Character specifying whether resampling should be done of groups,
observations within groups, or both. Currently only one level of grouping
is allowed.}

\item{parallel}{The type of parallel operation to be used (if any).}

\item{ncpus}{integer: number of processes to be used in parallel operation:
typically one would choose this to be the number of available CPUs.}
}
\description{
bootMerCases
}
