% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_code_generation.R
\name{generate_slurm_file}
\alias{generate_slurm_file}
\title{generate_slurm_file}
\usage{
generate_slurm_file(bash_out_dir, nreps, niter, mc.cores,
  simtype = "permutation", J = 100, n_j = 4, a = 0, b = 0,
  c_p = 0, theta_ab = 0.2, re.form = NULL, permtype = "within",
  boottype = "parametric", optimizer = "bobyqa",
  job_name = "permediatr", job_time = "2-00:00:00", job_mem = "5G",
  partition = "ncf_holy", email_address = NULL, save_dir = "./",
  fname_prefix = "pmjob")
}
\arguments{
\item{re.form}{}
}
\description{
generate_slurm_file
}
