% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{indirectEffect}
\alias{indirectEffect}
\title{indirectEffect}
\usage{
indirectEffect(x)
}
\arguments{
\item{x}{a permediatrMod object}
}
\value{
an indirectEffect object, with elements ab, the indirect effect
  estiamte, warnings, a vector of warnings, and singular, a vector of
  logicals indicating if convergence was singular.
}
\description{
Takes a permediatrMod or permediatrPermMod object and extracts the estimate
of the indirect effect.
}
