% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{lmemm_within.perm}
\alias{lmemm_within.perm}
\title{lmemm_within.perm}
\usage{
lmemm_within.perm(...)
}
\arguments{
\item{...}{Passed on to the permediatrPermMod constructor}
}
\value{
indirectEffect object.
}
\description{
Variables should be centered appropriately (i.e., within-person)
}
