% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_bootFUN}
\alias{make_bootFUN}
\title{make_bootFUN}
\usage{
make_bootFUN(x, path = c("a", "b"), is.null.model = F)
}
\arguments{
\item{x}{A permediatrMod object with the full mediation specification.}

\item{path}{Which path is being bootstrapped, the 'a' path, or the 'b' path?}

\item{is.null.model}{Should the boot FUN expect to get a null model without
the effect of interest? Set this to true of you're doing parametric or
semiparametric bootstrapping for the nil null distribution. The
\code{\link{make_null_formula}} should be used to create the null model to
be the basis for (semi-)parametric simulation.}
}
\description{
make_bootFUN
}
