% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_null_formula}
\alias{make_null_formula}
\title{make_null_formula}
\usage{
make_null_formula(model, name, random)
}
\arguments{
\item{model}{A specific model from a permediatrMod object.}

\item{name}{The variable name to be excluded from the model.}

\item{random}{Whether the variable name is also included in the random
effects part of the model.}
}
\description{
make_null_formula
}
