% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{make_null_model}
\alias{make_null_model}
\title{make_null_model}
\usage{
make_null_model(x, path = c("a", "b"))
}
\arguments{
\item{x}{A permediatrMod object with the full mediation specification.}

\item{path}{For which path is the null model being generated, the 'a' path, or the 'b' path?}
}
\description{
make_null_model
}
