% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{permediatrMod}
\alias{permediatrMod}
\title{permediatrMod}
\usage{
permediatrMod(data, y.name, x.name, m_b.name, m_a.name, group.id,
  covariates.y = NULL, covariates.m = NULL, random.a = T,
  random.b = T, random.c_p = T, optimizer = "nloptwrap",
  lmeropts = list())
}
\arguments{
\item{re.form.m}{}
}
\description{
permediatrMod
}
