% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{permediatrPermMod}
\alias{permediatrPermMod}
\title{permediatrPermMod}
\usage{
permediatrPermMod(x, indices.y = NULL, indices.m = NULL,
  re.form.y = NULL, re.form.m = NULL)
}
\arguments{
\item{x}{a permediatrMod object}

\item{indices.y}{indices to use to permute residuals for model.y}

\item{indices.m}{indices to use to permute residuals for model.m}

\item{re.form.y}{random effects form for y}

\item{re.form.m}{random effects form for m}
}
\value{
an object of class permediatrPermMod
}
\description{
permediatrPermMod
}
