% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{permute_ab}
\alias{permute_ab}
\title{permute_ab}
\usage{
permute_ab(x, nperms, mc.cores, re.form = NULL, permtype = "within")
}
\arguments{
\item{x}{a permediatrMod object}

\item{permtype}{}
}
\description{
permute_ab
}
