% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{permute_between_within}
\alias{permute_between_within}
\title{permute_between_within}
\usage{
permute_between_within(n, data, group.id, series = F)
}
\arguments{
\item{series}{}
}
\description{
permute_between_within
}
