% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_bootstrap_simulation}
\alias{run_bootstrap_simulation}
\title{run_bootstrap_simulation}
\usage{
run_bootstrap_simulation(nreps, nperms, mc.cores, J = 100, n_j = 4,
  a = 0, b = 0, c_p = 0, theta_ab = 0.2, optimizer = "bobyqa",
  re.form = NULL, permtype = "within")
}
\arguments{
\item{permtype}{}
}
\description{
run_bootstrap_simulation
}
