% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_permutation_simulation}
\alias{run_permutation_simulation}
\title{run_permutation_simulation}
\usage{
run_permutation_simulation(nreps, niter, mc.cores, ...)
}
\arguments{
\item{nreps}{Number of repetitions using new data.}

\item{niter}{For each simulated data set, number of permutation or bootstrap iterations.}

\item{mc.cores}{Number of cores to use.}

\item{...}{To be passed to \code{\link{run_simulation}}}
}
\description{
Included for backward compatibility.
}
