% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_mediation_data}
\alias{simulate_mediation_data}
\title{Title}
\usage{
simulate_mediation_data(J = 100, n_j = 4, a = 0, b = 0, c_p = 0,
  theta_ab = 0.2)
}
\arguments{
\item{theta_ab}{}
}
\description{
Title
}
