% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{summary.permediatrBoot}
\alias{summary.permediatrBoot}
\title{summary.permediatrBoot}
\usage{
\method{summary}{permediatrBoot}(x, which = "ab", verbose = F)
}
\arguments{
\item{x}{An object of class permediatrBoot.}

\item{verbose}{}
}
\value{
some p-values and other information.
}
\description{
summary.permediatrBoot
}
