% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediations.R
\name{indirect_within.lme4}
\alias{indirect_within.lme4}
\title{indirect_within.lme4}
\usage{
indirect_within.lme4(x, indices.y = NULL, indices.m = NULL,
  re.form.y = NULL, re.form.m = NULL, ...)
}
\arguments{
\item{data}{Data frame. Should not have missing values.}

\item{indices}{For permuting. Should be the same length as dim(data)[1]}

\item{y.name}{y variable name.}

\item{x.name}{x variable name (should be within-person centered).}

\item{m_b.name}{mediator variable name (within-person centered for computing b path).}

\item{m_a.name}{mediator variable name (raw for computing a path).}

\item{group.id}{grouping variable for random effects.}

\item{covariates.y}{any covariates to include in regression of y.}

\item{covariates.m}{any covariates to include in regression of m.}

\item{random.a}{include random effect for m on x (a path).}

\item{random.b}{include random effect for y on m (b path).}

\item{random.c_p}{include random effect for y on x (c' path).}

\item{optimizer}{}
}
\value{
indirect and total effect.
}
\description{
Variables should be centered appropriately (i.e., within-person)
}
